+++ neon/builtin/contains_test.go
@@ -17,8 +17,7 @@ func TestContains(t *testing.T) {
 	if contains([]interface{}{"foo", "bar"}, "spam") {
 		t.Errorf("Error builtin contains")
 	}
-	var content interface{}
-	content = []string{"foo", "bar"}
+	content := []string{"foo", "bar"}
 	if !contains(content, "bar") {
 		t.Errorf("Error builtin contains")
 	}