+++ neon/main_test.go
@@ -24,7 +24,9 @@ foo: bars`
 	if err := os.WriteFile(file, []byte(source), 0644); err != nil {
 		t.Errorf("Error writing configuration file")
 	}
-	defer os.Remove(file)
+	defer func() {
+		_ = os.Remove(file)
+	}()
 	configuration, err := ParseConfiguration(file)
 	if err != nil {
 		t.Errorf("Error parsing configuration")
@@ -49,7 +51,9 @@ func TestParseConfigurationError(t *testing.T) {
 	if err := os.WriteFile(file, []byte(source), 0644); err != nil {
 		t.Errorf("Error writing configuration file")
 	}
-	defer os.Remove(file)
+	defer func() {
+		_ = os.Remove(file)
+	}()
 	_, err := ParseConfiguration(file)
 	if err == nil || err.Error() != "yaml: line 1: did not find expected '-' indicator" {
 		t.Errorf("Error parsing configuration: %v", err)
@@ -71,7 +75,9 @@ foo: bars`
 	if err := os.WriteFile(file, []byte(source), 0644); err != nil {
 		t.Errorf("Error writing configuration file")
 	}
-	defer os.Remove(file)
+	defer func() {
+		_ = os.Remove(file)
+	}()
 	_, err := LoadConfiguration(file)
 	if err != nil {
 		t.Errorf("Error parsing configuration")
@@ -84,10 +90,10 @@ foo: bars`
 func TestParseCommandLine(t *testing.T) {
 	os.Args = []string{"cmd", "-file", "file", "-info", "-version", "-props", "{foo: bar, spam: eggs}",
 		"-time", "-tasks", "-task", "task", "-targets", "-builtins", "-builtin", "builtin", "-tree",
-		"-tasks-ref", "-builtins-ref", "-install", "install", "-repo", "repo", "-update", "-grey",
+		"-tasks-ref", "-builtins-ref", "-install", "install", "-repo", "repo", "-update", "-batch", "-grey",
 		"-template", "template", "-templates", "-themes", "-theme", "test", "-parents", "target1", "target2"}
 	file, info, version, props, timeit, tasks, task, targs, builtins, builtin, tree, tasksRef, builtinsRef,
-		install, repo, update, grey, template, templates, parents, theme, themes, targets := ParseCommandLine()
+		install, repo, update, batch, grey, template, templates, parents, theme, themes, targets := ParseCommandLine()
 	Assert(file, "file", t)
 	Assert(info, true, t)
 	Assert(version, true, t)
@@ -104,6 +110,7 @@ func TestParseCommandLine(t *testing.T) {
 	Assert(install, "install", t)
 	Assert(repo, "repo", t)
 	Assert(update, true, t)
+	Assert(batch, true, t)
 	Assert(grey, true, t)
 	Assert(template, "template", t)
 	Assert(templates, true, t)