+++ neon/task/call.go
@@ -1,8 +1,11 @@
 package task
 
 import (
-	"github.com/c4s4/neon/neon/build"
+	"fmt"
+	"os"
 	"reflect"
+
+	"github.com/c4s4/neon/neon/build"
 )
 
 func init() {
@@ -29,6 +32,10 @@ type callArgs struct {
 
 func call(context *build.Context, args interface{}) error {
 	params := args.(callArgs)
+	dir, err := os.Getwd()
+	if err != nil {
+		return fmt.Errorf("getting current directory: %v", err)
+	}
 	for _, target := range params.Call {
 		stack := context.Stack.Copy()
 		context.MessageArgs("Calling target '%s'", target)
@@ -37,6 +44,9 @@ func call(context *build.Context, args interface{}) error {
 			return err
 		}
 		context.Stack = stack
+		if err := os.Chdir(dir); err != nil {
+			return fmt.Errorf("changing to build directory: %v", err)
+		}
 	}
 	return nil
 }