+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # NeON
 
+## 2025-07-17: 1.15.0
+
+- Added `pause` task
+
+## 2025-05-06: 1.14.1
+
+- Fix dotenv loading
+
+## 2025-05-06: 1.14.0
+
+- Added `properties` key to `neon` task
+- Added `write()` builtin to write text into a file
+
+## 2025-04-14: 1.13.1
+
+- Fixed code warnings
+- Fix start task
+
+## 2025-03-13: 1.13.0
+
+- Added `jsonindent()` builtin
+
+## 2025-03-06: 1.12.0
+
+- Added `env` field to `shell` task to set environment
+
+## 2025-02-18: 1.11.1
+
+- Added `-batch` option to force plugin update with `-update` option
+
 ## 2025-02-13: 1.11.0
 
 - Environment variables defined in build file are defined everywhere in it, not only in executed commands